+++ projects/query-search/src/lib/components/fields/name-field/name-field.component.ts
@@ -1,30 +1,53 @@
-import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
+import {Component, EventEmitter, Input, Output} from '@angular/core';
 import {QuerySearchService} from "../../../query-search.service";
-import {Observable} from "rxjs";
+import {BehaviorSubject, Observable} from "rxjs";
 import {QueryField} from "../../../models";
+import {map, startWith, switchMap} from "rxjs/operators";
 
 @Component({
   selector: 'name-field',
   templateUrl: './name-field.component.html',
   styleUrls: ['./name-field.component.scss']
 })
-export class NameFieldComponent implements OnInit {
+export class NameFieldComponent {
   @Output()
   selectedFieldChange = new EventEmitter<QueryField>(true);
 
   @Input()
   selectedField: QueryField;
 
-  fields: Observable<QueryField[]>;
+  allFields: Observable<QueryField[]>;
+  visibleFields: Observable<QueryField[]>;
+  searchValue: BehaviorSubject<string> = new BehaviorSubject<string>('');
+  nameFieldTrigger: string;
 
   constructor(private querySearchService: QuerySearchService) {
-    this.fields = querySearchService.fields;
-  }
-
-  ngOnInit(): void {
+    this.allFields = querySearchService.fields;
+    this.visibleFields = this.searchValue.pipe(
+      startWith(null),
+      switchMap(searchValue => {
+        return this.allFields.pipe(
+          map(fields => {
+            if (!!searchValue && searchValue.trim().length > 0) {
+              return fields.filter(f => f.name.toLowerCase().includes(searchValue))
+            }
+            return fields;
+          })
+        )
+      })
+    )
   }
 
   get formFieldAppearance() {
     return this.querySearchService.formFieldAppearance;
+  }
+
+  searchValueChanged(searchValue: string) {
+    this.searchValue.next(searchValue);
+  }
+
+  fieldChanged(event: QueryField) {
+    this.selectedFieldChange.emit(event);
+    this.nameFieldTrigger = `<span class="field-name">${event.label || event.name}</span><span class="field-suffix">Name</span>`;
   }
 }