+++ projects/query-search/src/lib/components/query-search-item/query-search-item.component.ts
@@ -1,5 +1,4 @@
 import {
-  AfterViewInit,
   ChangeDetectionStrategy, ChangeDetectorRef,
   Component,
   ElementRef,
@@ -9,12 +8,12 @@ import {
   QueryList,
   ViewChildren
 } from '@angular/core';
-import {ProvidedValue, QueryField, QueryItem} from "../../models";
+import {QueryField, QueryItem} from "../../models";
 import {QuerySearchService} from "../../query-search.service";
-import {BehaviorSubject, fromEvent, Observable, of, Subscription} from "rxjs";
+import {Observable} from "rxjs";
 import {DateAdapter} from "@angular/material/core";
 import {CustomDateAdapter} from "../../adapters";
-import {distinctUntilChanged, filter, map, shareReplay, tap} from "rxjs/operators";
+
 
 @Component({
   selector: 'query-search-item',
@@ -25,7 +24,7 @@ import {distinctUntilChanged, filter, map, shareReplay, tap} from "rxjs/operator
   ],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class QuerySearchItemComponent implements AfterViewInit {
+export class QuerySearchItemComponent {
 
   @Input()
   set item(newValue: QueryItem) {
@@ -49,58 +48,21 @@ export class QuerySearchItemComponent implements AfterViewInit {
   @HostBinding('class.double-height')
   doubleHeight = false;
 
-  totalValues: BehaviorSubject<number> = new BehaviorSubject<number>(0);
-  allValues: any[] | ProvidedValue[] = [];
-  visibleValues: any[] | ProvidedValue[] = [];
-  searchValue: string = '';
-  selectedValues: any[] = [];
   leftDate: Date;
   rightDate: Date;
   operators: string[];
   fields: Observable<QueryField[]>;
   selectedField: QueryField;
-  $loading: BehaviorSubject<boolean> = new BehaviorSubject<boolean>(true);
-  operatorStyle = `
-  position: absolute;
-   right: 1em;
-   color: gray
-   `;
-
-  values: Observable<any[]>;
   showBetweenDateFields = false;
 
   private _currentOperator: string;
   private _item: QueryItem;
-  private _valuesObservable: Observable<any[]>;
-  private _fieldValueSubscription: Subscription;
 
   constructor(private querySearchService: QuerySearchService,
               private changeDetectorRef: ChangeDetectorRef,
               private dateAdapter: DateAdapter<any>) {
     this.operators = querySearchService.operators;
     this.fields = querySearchService.fields;
-  }
-
-
-  ngAfterViewInit() {
-    // Listen for input changes on an autocomplete input
-    this.valueInputs.changes.subscribe((changes: QueryList<ElementRef>) => {
-      if (!!changes && !!changes.first) {
-        this.setupFieldValueListener(changes.first);
-      }
-    });
-
-    // Remove the stupid ripple
-    this.searchInputs.changes.subscribe((changes: QueryList<ElementRef>) => {
-      if (!!changes && !!changes.first) {
-        const matOption = changes.first.nativeElement.parentElement.parentElement.parentElement
-        const matRipple = matOption.querySelector('.mat-ripple');
-
-        if (!!matRipple) {
-          matRipple.remove();
-        }
-      }
-    })
   }
 
   remove() {
@@ -113,11 +75,8 @@ export class QuerySearchItemComponent implements AfterViewInit {
     this.item.fieldName = field.name;
     this.item.type = field.type;
     this.querySearchService.log('Field selected', field);
+
     this.updateDateFormat();
-    this.checkForObservable();
-    this.item.value = null;
-    this.selectedValues = [];
-    this.values = this.getValues();
   }
 
   get hasValues(): boolean {
@@ -168,44 +127,6 @@ export class QuerySearchItemComponent implements AfterViewInit {
     return false;
   }
 
-  getValues(): Observable<any[] | ProvidedValue[]> {
-    if (!!this.selectedField && !!this.selectedField.values) {
-      if (this.selectedField.values instanceof Observable) {
-        return this._valuesObservable.pipe(
-          tap(values => {
-            this.updateValues(values);
-            this.totalValues.next(values.length);
-          })
-        )
-      } else if (this.selectedField.values.length > 0) {
-        this.updateValues(this.selectedField.values);
-        this.totalValues.next(this.selectedField.values.length);
-        return of(this.selectedField.values)
-      }
-    }
-
-    if (this.selectedField.type === 'boolean') {
-      const booleanValues = [
-        {
-          value: 'true',
-          displayValue: 'True'
-        },
-        {
-          value: 'false',
-          displayValue: 'False'
-        }
-      ];
-
-      this.totalValues.next(2);
-      this.updateValues(booleanValues);
-      return of(booleanValues).pipe(
-        shareReplay(1)
-      )
-    }
-
-    this.totalValues.next(0);
-    return of([]);
-  }
 
   dateUpdated(newDate: Date) {
     this.querySearchService.log('Date updated', newDate);
@@ -216,42 +137,6 @@ export class QuerySearchItemComponent implements AfterViewInit {
     this.item.value = [this.leftDate, this.rightDate];
     this.querySearchService.log('Date updated', this.leftDate, this.rightDate);
     this.querySearchService.log('Formatted date', this.item.value);
-  }
-
-  valueSelected(value: any | ProvidedValue) {
-    if (value.hasOwnProperty('displayValue') && value.hasOwnProperty('value')) {
-      return this.selectedValues.includes(value.value);
-    }
-
-    return this.selectedValues.includes(value);
-  }
-
-  optionClicked(event: Event, value: any) {
-    event.stopPropagation();
-    this.toggleSelection(value);
-  }
-
-  searchOptionClicked(event: Event) {
-    event.stopImmediatePropagation();
-  }
-
-  toggleSelection(value: any | ProvidedValue) {
-    if (!this.valueSelected(value)) {
-      if (value.hasOwnProperty('displayValue') && value.hasOwnProperty('value')) {
-        this.selectedValues.push(value.value);
-      } else {
-        this.selectedValues.push(value)
-      }
-    } else {
-      if (value.hasOwnProperty('displayValue') && value.hasOwnProperty('value')) {
-        this.selectedValues = this.selectedValues.filter(v => v !== value.value);
-      } else {
-        this.selectedValues = this.selectedValues.filter(v => v !== value);
-      }
-    }
-
-    this.item.value = this.selectedValues.join(',');
-    this.changeDetectorRef.detectChanges();
   }
 
   get formFieldAppearance() {
@@ -275,8 +160,6 @@ export class QuerySearchItemComponent implements AfterViewInit {
   }
 
   operatorSelected(operator: any) {
-    setTimeout(this.setupFieldValueListener, 300);
-    this.selectedValues = [];
     this._currentOperator = operator;
     this.item.value = null;
     this.querySearchService.log('Clearing selected values for', this);
@@ -290,35 +173,6 @@ export class QuerySearchItemComponent implements AfterViewInit {
     }
   }
 
-  searchValues(event: Event) {
-    event.stopImmediatePropagation();
-  }
-
-  searchValueChanged(partialValue: string) {
-    if (partialValue.trim().length) {
-      this.visibleValues = this.allValues.filter((value: any | ProvidedValue) => {
-        const lowerValue = partialValue.toLowerCase();
-        if (value.hasOwnProperty('displayValue')) {
-          if (value.hasOwnProperty('description') && value.description) {
-            return value.value.toLowerCase().includes(lowerValue) || value.displayValue.toLowerCase().includes(lowerValue) || value.description.toLowerCase().includes(lowerValue)
-          }
-
-          return value.value.toLowerCase().includes(lowerValue) || value.displayValue.toLowerCase().includes(lowerValue)
-        }
-
-        return `${value}`.includes(lowerValue);
-      })
-    } else {
-      this.visibleValues = this.allValues;
-    }
-  }
-
-  clearSearch() {
-    this.searchValueChanged('');
-    if (!!this.searchInputs.first) {
-      this.searchInputs.first.nativeElement.value = '';
-    }
-  }
 
   private loadFieldFromItem() {
     if (!!this.item && !!this.item.fieldName && !this.selectedField) {
@@ -329,17 +183,8 @@ export class QuerySearchItemComponent implements AfterViewInit {
           this.item.type = field.type;
           this.querySearchService.log('Field loaded', field);
           this.updateDateFormat();
-          this.checkForObservable();
         }
       })
-    }
-  }
-
-  private checkForObservable() {
-    if (!!this.selectedField && !!this.selectedField.values && this.selectedField.values instanceof Observable) {
-      this._valuesObservable = this.selectedField.values.pipe(
-        tap(() => this.$loading.next(false))
-      )
     }
   }
 
@@ -347,40 +192,5 @@ export class QuerySearchItemComponent implements AfterViewInit {
     if (!!this.selectedField && !!this.selectedField.format) {
       (this.dateAdapter as CustomDateAdapter).setFormat(this.selectedField.format);
     }
-  }
-
-  private setupFieldValueListener(field: ElementRef) {
-    if (this._fieldValueSubscription) {
-      this._fieldValueSubscription.unsubscribe();
-    }
-
-    if (!!field) {
-      this._fieldValueSubscription = fromEvent(field.nativeElement, 'keyup').pipe(
-        filter(Boolean),
-        distinctUntilChanged(),
-        map(() => field.nativeElement.value),
-        tap(inputValue => this.processValuesSelected(inputValue)),
-      ).subscribe(inputValue => {
-        this.querySearchService.valueFieldChanged(this.selectedField, inputValue);
-      });
-    }
-  }
-
-  private processValuesSelected(inputValue: string) {
-    if (!!inputValue && inputValue.length) {
-      this.selectedValues = inputValue.split(',');
-    } else {
-      this.selectedValues = [];
-    }
-
-    this.changeDetectorRef.detectChanges();
-  }
-
-  private updateValues(values: any[] | ProvidedValue[]) {
-    if (this.searchValue.length === 0) {
-      this.visibleValues = values;
-    }
-
-    this.allValues = values;
   }
 }